pyproject.toml
[project]
name = "scratchgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flowchat>=1.4.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "scratchattach>=2.1.11",
]
